requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.2
pandas==2.1.4
numpy==1.24.4
scikit-learn==1.3.2
plotly==5.17.0
openpyxl==3.1.2
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.0
xgboost==2.0.3
seaborn==0.13.0
matplotlib==3.8.2
reportlab==4.0.8
python-pptx==0.6.23
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
duckdb==0.9.2
asyncio==3.4.3
aiofiles==23.2.0
Pillow==10.1.0